优化：拆分平台监听到独立文件，提升代码可读性
