+++ data/抖音礼物价格表.json
@@ -1,4 +1,5 @@
 {
+    "点亮粉丝团": 1,
     "小心心": 1,
     "玫瑰": 1,
     "抖音": 1,
